More Semantic Modifications to the HTML Structure
Upon further inspection, there were several issues with the current HTML structure.

These included an id attribute that was mistakingly put in the footer tag, as well as a section tag structure that was not conducive to the design.

To solve this, the document was reformatted. Additonally, an Image tag placeholder was put after the submit button near the top of the page.