Restructuring of the Header and First Section in the HTML
There was a problem where the HTML document format would not support the mockup styling.

The total text area needed to be at 80% width, however the background color needed to fill the screen's width.

I solved this by adding div containers in order to separately style the child elements of the elements with background colors.