Modified Layout of item with the class of placeholder
The original p.placeholder wouldn't properly style according to the project's specifications.

In order to solve this, I wrapped the p element in a div and gave the div the class instead.

In terms of stylings, I grayed out the background for the div and also used padding on the div to center the p text.